refactor: api 엔드포인트 경로와 view 경로를 명세에 맞춰 수정, 일부 기능 버그 수정

+++ src/main/java/io/cavia/trader/module/member/service/MemberService.java
@@ -33,8 +33,8 @@ public interface MemberService {
 
     void createMember(Member member);
 
-    List<UserRankingDto> findAllOrderByCash(int limit , int offset);
+    List<UserRankingDto> findAllOrderByCash(Long limit , Long offset);
 
-    List<UserRankingDto> findAllOrderByTotalScore(int limit , int offset);
+    List<UserRankingDto> findAllOrderByTotalScore(Long limit , Long offset);
 
 }
